requirements.txt
accelerate==1.6.0
antlr4-python3-runtime==4.9.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
grad-cam==1.5.5
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
peft==0.15.1
pillow==11.2.1
portalocker==3.1.1
psutil==7.0.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
-e git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4#egg=SAM_2
sam2==1.1.0
scikit-learn==1.6.1
scipy==1.15.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
ttach==0.0.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
